Modifying the scheduler to account for RDataFrame Defines being truly lazy.
An RDataFrame Define is truly lazy. This means that the variable is not calculated at the point in the computational graph that it is declared but rather where it is first used. This means that there is no point being too clever with scheduling variables because they are already evaluated only at the absolutely latest possible point.
This commit therefore changes the scheduling behaviour. All define calls are now scheduled at the start of the graph (in an order that ensures that no variable is declared before its inputs). The scheduling of filters and fills still considers the variables as dependencies. This serves to ensure that the correct *filter* dependencies of each variable are transferred to any filters that use them.

Consider a scenario where a region is defined by a filter 'X'. This filter has an input variable 'a' which itself depends on a filter 'Y'. Previously this would have been scheduled as
Y -> a -> X.
Now it will be scheduled as
a -> Y -> X.
By placing Y before X, the scheduler ensures that a is still never evaluated on an event that doesn't pass Y!